backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
pydantic
python-dotenv
supabase
google-generativeai
